refactor: 기존 ShareTicket 폴더명 DB 테이블명 변경에 따라 AttendeeForm으로 변경
