+++ src/main/java/com/fairing/fairplay/attendee/service/AttendeeService.java
@@ -15,8 +15,8 @@ import com.fairing.fairplay.qr.dto.QrTicketEmailTodayRequestDto;
 import com.fairing.fairplay.qr.service.QrTicketService;
 import com.fairing.fairplay.reservation.entity.Reservation;
 import com.fairing.fairplay.reservation.repository.ReservationRepository;
-import com.fairing.fairplay.shareticket.entity.ShareTicket;
-import com.fairing.fairplay.shareticket.service.ShareTicketService;
+import com.fairing.fairplay.attendeeform.entity.ShareTicket;
+import com.fairing.fairplay.attendeeform.service.ShareTicketService;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Objects;
