+++ src/main/java/com/fairing/fairplay/attendeeform/controller/ShareTicketController.java
@@ -1,12 +1,10 @@
-package com.fairing.fairplay.shareticket.controller;
+package com.fairing.fairplay.attendeeform.controller;
 
 import com.fairing.fairplay.core.security.CustomUserDetails;
-import com.fairing.fairplay.shareticket.dto.ShareTicketInfoResponseDto;
-import com.fairing.fairplay.shareticket.dto.ShareTicketSaveRequestDto;
-import com.fairing.fairplay.shareticket.dto.ShareTicketSaveResponseDto;
-import com.fairing.fairplay.shareticket.dto.TokenResponseDto;
-import com.fairing.fairplay.shareticket.service.ShareTicketAttendeeService;
-import com.fairing.fairplay.shareticket.service.ShareTicketService;
+import com.fairing.fairplay.attendeeform.dto.ShareTicketInfoResponseDto;
+import com.fairing.fairplay.attendeeform.dto.TokenResponseDto;
+import com.fairing.fairplay.attendeeform.service.ShareTicketAttendeeService;
+import com.fairing.fairplay.attendeeform.service.ShareTicketService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -14,8 +12,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
