+++ src/main/java/com/fairing/fairplay/attendeeform/dto/ShareTicketInfoResponseDto.java
@@ -1,4 +1,4 @@
-package com.fairing.fairplay.shareticket.dto;
+package com.fairing.fairplay.attendeeform.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class TokenResponseDto {
+public class ShareTicketInfoResponseDto {
 
-  private String token;
+  private Long formId;
+  private Long eventId;
+  private String eventName;
 }
