+++ src/main/java/com/fairing/fairplay/attendeeform/dto/ShareTicketSaveRequestDto.java
@@ -1,4 +1,4 @@
-package com.fairing.fairplay.shareticket.dto;
+package com.fairing.fairplay.attendeeform.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
