+++ src/main/java/com/fairing/fairplay/attendeeform/dto/ShareTicketSaveResponseDto.java
@@ -1,4 +1,4 @@
-package com.fairing.fairplay.shareticket.dto;
+package com.fairing.fairplay.attendeeform.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,9 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ShareTicketInfoResponseDto {
+public class ShareTicketSaveResponseDto {
 
-  private Long formId;
-  private Long eventId;
-  private String eventName;
+  private Long reservationId;
+  private String token;
 }
