+++ src/main/java/com/fairing/fairplay/payment/service/PaymentService.java
@@ -25,9 +25,9 @@ import com.fairing.fairplay.reservation.entity.Reservation;
 import com.fairing.fairplay.reservation.repository.ReservationRepository;
 import com.fairing.fairplay.reservation.repository.ReservationStatusCodeRepository;
 import com.fairing.fairplay.reservation.service.ReservationService;
-import com.fairing.fairplay.shareticket.dto.ShareTicketSaveRequestDto;
-import com.fairing.fairplay.shareticket.dto.ShareTicketSaveResponseDto;
-import com.fairing.fairplay.shareticket.service.ShareTicketAttendeeService;
+import com.fairing.fairplay.attendeeform.dto.ShareTicketSaveRequestDto;
+import com.fairing.fairplay.attendeeform.dto.ShareTicketSaveResponseDto;
+import com.fairing.fairplay.attendeeform.service.ShareTicketAttendeeService;
 import com.fairing.fairplay.ticket.entity.EventSchedule;
 import com.fairing.fairplay.ticket.entity.Ticket;
 import com.fairing.fairplay.ticket.repository.EventScheduleRepository;
